Update contour to 1.2.1

+++ pkg/reconciler/contour/resources/httpproxy.go
@@ -82,7 +82,7 @@ func MakeHTTPProxies(ctx context.Context, ing *v1alpha1.Ingress, serviceToProtoc
 			var retry *v1.RetryPolicy
 			if path.Retries != nil && path.Retries.Attempts > 0 {
 				retry = &v1.RetryPolicy{
-					NumRetries:    uint32(path.Retries.Attempts),
+					NumRetries:    int64(path.Retries.Attempts),
 					PerTryTimeout: path.Retries.PerTryTimeout.Duration.String(),
 				}
 			}
@@ -126,7 +126,7 @@ func MakeHTTPProxies(ctx context.Context, ing *v1alpha1.Ingress, serviceToProtoc
 				svcs = append(svcs, v1.Service{
 					Name:                 split.ServiceName,
 					Port:                 split.ServicePort.IntValue(),
-					Weight:               uint32(split.Percent),
+					Weight:               int64(split.Percent),
 					RequestHeadersPolicy: postSplitHeaders,
 					Protocol:             protocol,
 				})
